pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-restful>=0.3.10",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
]
